TPT-4456: Add Linode test with RDMA interfaces (#2452)
* add test and hidden

* improve test

* add update rdma test

* fix lint

* update tool dep

* nit

* fix lint

* update to linodego 2.6.0

* fix tflint

* go mod tidy

+++ linode/linodeinterface/framework_resource.go
@@ -219,6 +219,23 @@ func (r *Resource) Delete(
 		return
 	}
 
+	// RDMA VPC interfaces cannot be deleted through the API; they are only removed
+	// when their parent Linode is deleted. Drop the resource from state without
+	// issuing a delete request so that removing it from configuration does not fail.
+	if !state.RDMAVPC.IsNull() && !state.RDMAVPC.IsUnknown() {
+		resp.Diagnostics.AddWarning(
+			"RDMA VPC Interface Not Deleted",
+			fmt.Sprintf(
+				"RDMA VPC interface %v cannot be deleted via the linode_interface resource. "+
+					"It has been removed from Terraform state, but it remains attached to Linode %v "+
+					"and is no longer managed by Terraform. RDMA VPC interfaces are only removed when "+
+					"their parent Linode is deleted; to manage this interface again, re-import it.",
+				id, linodeID,
+			),
+		)
+		return
+	}
+
 	tflog.Debug(ctx, "client.DeleteInterface(...)")
 	err := client.DeleteInterface(ctx, linodeID, id)
 	if err != nil {
